// src/app/enums/Supply/PublicationType.ts
/* This Enum defines the possible publication types that can be used in the project. */
export enum PUBLICATION_TYPE {
    WEBSITE = 'website',
    MOBILE_APPLICATION = 'mobile_application',
    IOS_APPLICATION = 'ios_application',
    ANDROID_APPLICATION = 'android_application',
    BOTH = 'both',
}